+++ packages/signers/meta-signers/src/wrapped-tx-creator.ts
@@ -1,6 +1,6 @@
 import { TxCreator } from "@polkadot-api/polkadot-signer"
 
-export type WrapTxCreator<T extends TxCreator<any>> = T & {
+export type WrapTxCreator<T extends TxCreator> = {
   publicKey: Uint8Array
   accountId: Uint8Array
-}
+} & T